pyproject.toml
[project]
name = "facevision-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deepface>=0.0.95",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "face-recognition>=1.3.0",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "tf-keras>=2.20.1",
]
